Do not attempt to clean up thumbnails on delete!
This has large implications, mostly that we might need a mechanism for
keeping track of where thumbnails are being stored, so that we can go
through and clean out thumbnails of deleted content. Seems important.
But not a thing that the core media object needs to know how to do!
Requires more thinking.