Change how exporting and thumbnailing are handled!
Before this change, illuminatus attempted to manage not only the
database of media (mostly, if not all, metadata), but also an on-disk
collection of associated thumbnail content. These thumbnails were
generated largely through the same code path as the "export" mechanism,
which was a difficult merger because the two are aimed at different
audiences (thumbnails not for external consumption, exporting more so).
Now illuminatus doesn't handle thumbnails at all; all generated content
is maintained via exporting to disk. This moves burden from illuminatus
to the consumers of the illuminatus metadata. At the moment the web
server is the canonical such consumer, so it will have to handle
maintaining the thumbnails itself. WIP.